requirements.txt
chromadb==1.0.7
langchain==0.3.25
langchain-core==0.3.58
langchain-text-splitters==0.3.8
langchain-openai==0.1.7
langchain-chroma==0.1.5
langchain-community==0.0.30
sentence-transformers==4.1.0
PyMuPDF==1.25.5
streamlit==1.45.0
python-dotenv==1.0.1
